# mfdc-nc/docs/swagger.yaml
definitions:
...
  model.Reload:
    properties:
      reload:
        type: string
    type: object
...
info:
  contact: {}
  description: Swagger API for Golang Project MFDC NC
  title: MFDC Number changer
  version: "1.0"
paths:
  /config/reload:
    get:
      consumes:
      - application/json
      description: Reloading API configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Reload'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Execute reload config
      tags:
      - Reload
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-MFDC-Key
    type: apiKey
swagger: "2.0"